LayoutTests/imported/w3c/web-platform-tests/focus/focus-contenteditable-element-in-iframe-scroll-into-view-expected.txt

focusable 1


FAIL Check contenteditable element in an iframe scroll into view on second focusing assert_approx_equals: scroll X is within +/- 1 of a element in an iframe expected 0 +/- 1 but got 1519
